# docs/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lottery Service API",
    "description": "API для сервиса лотереи MEPHI Kotlin Band",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/tickets": {
      "post": {
        "tags": [
          "Tickets"
        ],
        "summary": "Создание билета",
        "description": "Создает билет без оплаты (устаревший метод)",
        "operationId": "createTicket",
        "deprecated": true,
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Билет успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные запроса"
          },
          "401": {
            "description": "Неавторизованный доступ"
          }
        }
      },
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Получение всех билетов пользователя",
        "description": "Возвращает список всех билетов текущего пользователя",
        "operationId": "getUserTickets",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список билетов успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ"
          }
        }
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Получение билета по ID",
        "description": "Возвращает информацию о билете по его ID",
        "operationId": "getTicketById",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID билета"
          }
        ],
        "responses": {
          "200": {
            "description": "Билет успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDto"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ"
          },
          "404": {
            "description": "Билет не найден"
          }
        }
      }
    },
    "/api/draws/{drawId}/tickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Получение билетов пользователя для тиража",
        "description": "Возвращает список билетов текущего пользователя для указанного тиража",
        "operationId": "getUserTicketsForDraw",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "drawId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID тиража"
          }
        ],
        "responses": {
          "200": {
            "description": "Список билетов успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Неавторизованный доступ"
          },
          "404": {
            "description": "Тираж не найден"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTicketRequest": {
        "type": "object",
        "required": [
          "drawId",
          "numbers"
        ],
        "properties": {
          "drawId": {
            "type": "string",
            "format": "uuid",
            "description": "ID тиража"
          },
          "numbers": {
            "type": "string",
            "description": "Выбранные номера в билете"
          }
        }
      },
      "TicketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Уникальный идентификатор билета"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "ID владельца билета"
          },
          "drawId": {
            "type": "string",
            "format": "uuid",
            "description": "ID тиража"
          },
          "numbers": {
            "type": "string",
            "description": "Выбранные номера в билете"
          },
          "status": {
            "type": "string",
            "description": "Статус билета",
            "enum": [
              "CREATED",
              "PAID",
              "PLAYED",
              "WON",
              "LOST"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Дата и время создания билета"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}